+++ src/components/form/b-select/b-select.ts
@@ -13,7 +13,7 @@
 
 import SyncPromise from 'core/promise/sync';
 
-import { derive } from 'core/functools/trait';
+import { derive } from 'components/traits';
 
 import Block, { setElementMod, removeElementMod, getElementSelector, element, elements } from 'components/friends/block';
 import DOM, { delegateElement } from 'components/friends/dom';
@@ -441,11 +441,7 @@ class bSelect extends iSelectProps implements iOpenToggle, iActiveItems {
 
 	protected override initBaseAPI(): void {
 		super.initBaseAPI();
-
-		const
-			i = this.instance;
-
-		this.normalizeItems = i.normalizeItems.bind(this);
+		this.normalizeItems = this.instance.normalizeItems.bind(this);
 	}
 
 	/** {@link iOpenToggle.initCloseHelpers} */