+++ src/core/component/meta/inherit.ts
@@ -6,9 +6,9 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-import { componentDecoratedKeys, PARENT } from 'core/component/const';
+import { PARENT } from 'core/component/const';
 
-import type { ModDeclVal, FieldWatcher } from 'core/component/interface';
+import type { ModDeclVal } from 'core/component/interface';
 import type { ComponentMeta } from 'core/component/meta/interface';
 
 /**
@@ -19,121 +19,35 @@ import type { ComponentMeta } from 'core/component/meta/interface';
  * @param parentMeta
  */
 export function inheritMeta(meta: ComponentMeta, parentMeta: ComponentMeta): ComponentMeta {
-	const decoratedKeys = componentDecoratedKeys[meta.componentName];
+	meta.tiedFields = {...parentMeta.tiedFields};
 
-	Object.assign(meta.tiedFields, parentMeta.tiedFields);
+	if (parentMeta.metaInitializers.size > 0) {
+		meta.metaInitializers = new Map(parentMeta.metaInitializers);
+	}
 
 	if (parentMeta.watchDependencies.size > 0) {
 		meta.watchDependencies = new Map(parentMeta.watchDependencies);
 	}
 
 	inheritParams(meta, parentMeta);
-	inheritProp(meta.props, parentMeta.props);
 
-	inheritField(meta.fields, parentMeta.fields);
-	inheritField(meta.systemFields, parentMeta.systemFields);
+	meta.props = Object.create(parentMeta.props);
+	meta.fields = Object.create(parentMeta.fields);
+	meta.systemFields = Object.create(parentMeta.systemFields);
 
-	inheritAccessors(meta.accessors, parentMeta.accessors);
-	inheritAccessors(meta.computedFields, parentMeta.computedFields);
+	meta.accessors = Object.create(parentMeta.accessors);
+	meta.computedFields = Object.create(parentMeta.computedFields);
+	meta.methods = Object.create(parentMeta.methods);
 
-	inheritMethods(meta.methods, parentMeta.methods);
+	meta.component.props = {...parentMeta.component.props};
+	meta.component.attrs = {...parentMeta.component.attrs};
+	meta.component.computed = {...parentMeta.component.computed};
 
 	if (meta.params.partial == null) {
 		inheritMods(meta, parentMeta);
 	}
 
 	return meta;
-
-	function inheritProp(current: ComponentMeta['props'], parent: ComponentMeta['props']) {
-		Object.entries(parent).forEach(([propName, parent]) => {
-			if (parent == null) {
-				return;
-			}
-
-			if (decoratedKeys == null || !decoratedKeys.has(propName)) {
-				current[propName] = parent;
-				return;
-			}
-
-			let watchers: CanUndef<Map<FieldWatcher['handler'], FieldWatcher>>;
-
-			parent.watchers?.forEach((watcher: FieldWatcher) => {
-				watchers ??= new Map();
-				watchers.set(watcher.handler, {...watcher});
-			});
-
-			current[propName] = {...parent, watchers};
-		});
-	}
-
-	function inheritField(current: ComponentMeta['fields'], parent: ComponentMeta['fields']) {
-		Object.entries(parent).forEach(([fieldName, parent]) => {
-			if (parent == null) {
-				return;
-			}
-
-			if (decoratedKeys == null || !decoratedKeys.has(fieldName)) {
-				current[fieldName] = parent;
-				return;
-			}
-
-			let
-				after: CanUndef<Set<string>>,
-				watchers: CanUndef<Map<FieldWatcher['handler'], FieldWatcher>>;
-
-			parent.watchers?.forEach((watcher: FieldWatcher) => {
-				watchers ??= new Map();
-				watchers.set(watcher.handler, {...watcher});
-			});
-
-			parent.after?.forEach((name: string) => {
-				after ??= new Set();
-				after.add(name);
-			});
-
-			current[fieldName] = {...parent, after, watchers};
-		});
-	}
-
-	function inheritAccessors(current: ComponentMeta['accessors'], parent: ComponentMeta['accessors']) {
-		Object.entries(parent).forEach(([accessorName, parent]) => {
-			current[accessorName] = {...parent!};
-		});
-	}
-
-	function inheritMethods(current: ComponentMeta['methods'], parent: ComponentMeta['methods']) {
-		Object.entries(parent).forEach(([methodName, parent]) => {
-			if (parent == null) {
-				return;
-			}
-
-			if (decoratedKeys == null || !decoratedKeys.has(methodName)) {
-				current[methodName] = {...parent};
-				return;
-			}
-
-			const
-				watchers = {},
-				hooks = {};
-
-			if (parent.watchers != null) {
-				Object.entries(parent.watchers).forEach(([key, val]) => {
-					watchers[key] = {...val};
-				});
-			}
-
-			if (parent.hooks != null) {
-				Object.entries(parent.hooks).forEach(([key, hook]) => {
-					hooks[key] = {
-						...hook,
-						after: Object.size(hook.after) > 0 ? new Set(hook.after) : undefined
-					};
-				});
-			}
-
-			current[methodName] = {...parent, watchers, hooks};
-		});
-	}
 }
 
 /**
@@ -175,7 +89,13 @@ export function inheritParams(meta: ComponentMeta, parentMeta: ComponentMeta): v
 export function inheritMods(meta: ComponentMeta, parentMeta: ComponentMeta): void {
 	const {mods} = meta;
 
-	Object.entries(parentMeta.mods).forEach(([modName, parentModValues]) => {
+	const keys = Object.keys(parentMeta.mods);
+
+	for (let i = 0; i < keys.length; i++) {
+		const
+			modName = keys[i],
+			parentModValues = parentMeta.mods[modName];
+
 		const
 			currentModValues = mods[modName],
 			forkedParentModValues = parentModValues?.slice() ?? [];
@@ -183,7 +103,7 @@ export function inheritMods(meta: ComponentMeta, parentMeta: ComponentMeta): voi
 		if (currentModValues != null) {
 			const values = Object.createDict<ModDeclVal>();
 
-			currentModValues.slice().forEach((modVal, i) => {
+			for (const [i, modVal] of currentModValues.slice().entries()) {
 				if (modVal !== PARENT) {
 					const modName = String(modVal);
 
@@ -191,14 +111,14 @@ export function inheritMods(meta: ComponentMeta, parentMeta: ComponentMeta): voi
 						values[modName] = <Exclude<typeof modVal, {}>>modVal;
 					}
 
-					return;
+					continue;
 				}
 
 				const hasDefault = currentModValues.some((el) => Object.isArray(el));
 
 				let appliedDefault = !hasDefault;
 
-				forkedParentModValues.forEach((modVal) => {
+				for (const modVal of forkedParentModValues) {
 					const modsName = String(modVal);
 
 					if (!(modsName in values)) {
@@ -209,10 +129,10 @@ export function inheritMods(meta: ComponentMeta, parentMeta: ComponentMeta): voi
 						forkedParentModValues[i] = modVal[0];
 						appliedDefault = true;
 					}
-				});
+				}
 
 				currentModValues.splice(i, 1, ...forkedParentModValues);
-			});
+			}
 
 			mods[modName] = <ModDeclVal[]>Object
 				.values(values)
@@ -221,5 +141,5 @@ export function inheritMods(meta: ComponentMeta, parentMeta: ComponentMeta): voi
 		} else if (!(modName in mods)) {
 			mods[modName] = forkedParentModValues;
 		}
-	});
+	}
 }